+++ corehq/warehouse/loaders/__init__.py
@@ -1,5 +1,6 @@
 from corehq.warehouse.loaders.domain import DomainStagingLoader, DomainDimLoader
 from corehq.warehouse.loaders.group import GroupStagingLoader, GroupDimLoader
+from corehq.warehouse.loaders.location import LocationStagingLoader, LocationDimLoader
 from corehq.warehouse.loaders.user import UserStagingLoader, UserDimLoader
 
 
@@ -11,5 +12,7 @@ def get_loader_by_slug(slug):
         UserDimLoader,
         GroupStagingLoader,
         GroupDimLoader,
+        LocationStagingLoader,
+        LocationDimLoader,
     ]
     return {cls.slug: cls for cls in loaders}[slug]
